refactor(app): extract open_file family to file_open_orchestrators (phase 5d)
Move open_file, open_file_no_focus, open_local_file,
open_file_with_encoding, reload_with_encoding,
open_file_large_encoding_confirmed, restore_global_file_state out of
buffer_management.rs into a new app/file_open_orchestrators.rs sibling.
~801 lines.

The biggest contiguous cluster in buffer_management.rs. Each opens a
file by coordinating: file-type detection, buffer creation/selection,
LSP registration, grammar parsing, per-file UI state restoration, and
focus assignment to the right split. Variants differ only in encoding
handling, focus behaviour, and "no file at this path yet" semantics.

buffer_management.rs goes from 1929 -> 1127 lines (-802).
2026 tests still pass.
